chore(Dafny): Move InternalResult into StandardLibrary(Internal) (#325)
