+++ AwsCryptographyPrimitives/runtimes/java/src/main/java/Signature/ECDSA.java
@@ -2,9 +2,9 @@ package Signature;
 
 import static Signature.SignatureAlgorithm.signatureAlgorithm;
 
-import Dafny.Aws.Cryptography.Primitives.Types.InternalResult;
 import Digest_Compile.ExternDigest;
 import Random_Compile.ExternRandom;
+import StandardLibraryInternal.InternalResult;
 import Wrappers_Compile.Result;
 import dafny.Array;
 import dafny.DafnySequence;
