+++ AwsCryptographyPrimitives/runtimes/java/src/main/java/Signature/PrivateKeyUtils.java
@@ -1,6 +1,6 @@
 package Signature;
 
-import Dafny.Aws.Cryptography.Primitives.Types.InternalResult;
+import StandardLibraryInternal.InternalResult;
 import Wrappers_Compile.Result;
 import dafny.Array;
 import dafny.DafnySequence;
